// src/api/orval/model/getApiV2VpnSessionsOverviewUsersPagedParams.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export type GetApiV2VpnSessionsOverviewUsersPagedParams = {
From: string;
To: string;
VpnServerId?: number;
ExternalId?: string;
DisplayName?: string;
/**
 * @minimum 1
 * @maximum 2147483647
 */
Page?: number;
/**
 * @minimum 1
 * @maximum 500
 */
PageSize?: number;
};