// src/api/orval/model/vpnServerClientsResponsesOverviewUsersV2Response.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { PagedVpnServerClientsDtoOverviewUserDto } from './pagedVpnServerClientsDtoOverviewUserDto';

export interface VpnServerClientsResponsesOverviewUsersV2Response {
  users?: PagedVpnServerClientsDtoOverviewUserDto;
}